Complete diffusion evaluation and robustness updates

+++ src/vrp_diffusion_quantum/metrics/matrix_metrics.py
@@ -145,7 +145,7 @@ def capacity_consistency_proxy(
         if component_demand <= capacity:
             feasible_customers += len(component)
 
-    return feasible_customers / n_customers
+    return float(feasible_customers / n_customers)
 
 
 @dataclass(frozen=True)
